src/main/java/com/example/catservice/service/CatFactService.java
package com.example.catservice.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Service;
import org.springframework.web.client.RestTemplate;

@Service
public class CatFactService {

    @Value("${catfact.url}")
    private String API_URL;

    public String getCatFact() {
        RestTemplate restTemplate = new RestTemplate();
        String response = restTemplate.getForObject(API_URL, String.class);
        return response;
    }

}