src/test/java/com/example/catservice/service/CatFactServiceTest.java
package com.example.catservice.service;

import com.github.tomakehurst.wiremock.WireMockServer;
import com.github.tomakehurst.wiremock.client.WireMock;
import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.web.client.TestRestTemplate;
import org.springframework.http.HttpStatus;
import org.springframework.test.context.TestPropertySource;

import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
import static com.github.tomakehurst.wiremock.client.WireMock.get;
import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@TestPropertySource(properties = "catfact.url = http://localhost:8080/fact")
class CatFactServiceTest {

    private WireMockServer wireMockServer;

    @Autowired
    private CatFactService catService;

    @BeforeEach
    public void setup() {
        wireMockServer = new WireMockServer(WireMockConfiguration.wireMockConfig().port(8080));
        wireMockServer.start();

        WireMock.configureFor("localhost", 8080);
        WireMock.stubFor(get(urlEqualTo("/fact"))
                .willReturn(aResponse()
                        .withStatus(HttpStatus.OK.value())
                        .withBody("This is a test cat fact.")));
    }

    @AfterEach
    public void tearDown() {
        wireMockServer.stop();
    }

    @Test
    public void testGetCatFact() {
        String catFact = catService.getCatFact();
        assertThat(catFact).isEqualTo("This is a test cat fact.");
    }
}